+++ src/twigextensions/ReadTimeTwigExtension.php
@@ -14,6 +14,7 @@ use lukeyouell\readtime\ReadTime;
 
 use Craft;
 use craft\helpers\DateTimeHelper;
+use craft\helpers\StringHelper;
 
 use yii\base\ErrorException;
 
@@ -43,41 +44,57 @@ class ReadTimeTwigExtension extends \Twig_Extension
 
     public function readTimeFunction($element, $showSeconds = true)
     {
-        $settings = ReadTime::$plugin->getSettings();
-        $wpm = $settings->wordsPerMinute;
         $totalSeconds = 0;
+        $vals = '';
 
         foreach ($element->getFieldLayout()->getFields() as $field) {
             try {
-                $fieldVal = $element->getFieldValue($field->handle);
+                // If field is a matrix then loop through fields in block
+                if ($field instanceof craft\fields\Matrix) {
+                    foreach($element->getFieldValue($field->handle)->all() as $block) {
+                        $blockFields = $block->getFieldLayout()->getFields();
 
-                $value = is_array($fieldVal) ? implode(' ', $fieldVal) : (string)$fieldVal;
-                $words = str_word_count(strip_tags($value));
-                $seconds = floor($words / $wpm * 60);
-
-                $totalSeconds = $totalSeconds + $seconds;
+                        foreach($blockFields as $blockField){
+                            $value = $block->getFieldValue($blockField->handle);
+                            $seconds = $this->valToSeconds($value);
+                            $totalSeconds = $totalSeconds + $seconds;
+                        }
+                    }
+                } else {
+                  $value = $element->getFieldValue($field->handle);
+                  $seconds = $this->valToSeconds($value);
+                  $totalSeconds = $totalSeconds + $seconds;
+                }
             } catch (ErrorException $e) {
                 continue;
             }
         }
 
-        $duration = DateTimeHelper::secondsToHumanTimeDuration($seconds, $showSeconds);
+        $duration = DateTimeHelper::secondsToHumanTimeDuration($totalSeconds, $showSeconds);
 
         return $duration;
     }
 
     public function readTimeFilter($value = null, $showSeconds = true)
     {
-        $settings = ReadTime::$plugin->getSettings();
-
-        $value = is_array($value) ? implode(' ', $value) : (string)$value;
-        $wpm = $settings->wordsPerMinute;
-
-        $words = str_word_count(strip_tags($value));
-        $seconds = floor($words / $wpm * 60);
-
+        $seconds = $this->valToSeconds($value);
         $duration = DateTimeHelper::secondsToHumanTimeDuration($seconds, $showSeconds);
 
         return $duration;
+    }
+
+    // Private Methods
+    // =========================================================================
+
+    private function valToSeconds($value)
+    {
+        $settings = ReadTime::$plugin->getSettings();
+        $wpm = $settings->wordsPerMinute;
+
+        $string = StringHelper::toString($value);
+        $wordCount = StringHelper::countWords($string);
+        $seconds = floor($wordCount / $wpm * 60);
+
+        return $seconds;
     }
 }