+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## 1.2.1 - 2018-06-19
+
+### Fixed
+- `readTime()` function now includes matrix fields when calculating the read time
+
 ## 1.2.0 - 2018-06-18
 
 ### Added
-- Calculate the read time of the whole entry based on it's field layout 
+- Calculate the read time of the whole entry based on it's field layout
 
 ## 1.1.0 - 2018-06-06